projetos/reconhecimento_optico_de_caracteres_tesseract/ocr/src/requirements.txt
# Frameworks e Bibliotecas
# Python 3.8
#accelerate>=0.26.0,<0.27.0
#arch
#boto3==1.18.0
#catboost>=1.2.7,<2.0
#click>=8.0.1,<9.0
#cython>=0.29.21,<0.30
#dask[dataframe]>=2021.9.1
#dask-ml>=2022.1.1,<2023.0
#emoji>=2.6.0
# fastparquet==0.5.0
#fasttext>=0.9.2,<1.0
#flask>=2.0,<3.0
#gensim>=4.0.1,<5.0
#imbalanced-learn>=0.8.0,<0.10
#ipython==7.34.0
#joblib>=1.0.1,<2.0
#kafka-python==2.0.2
#keras>=2.8.0,<2.9.0
#langdetect
#lime==0.2.0.1
#lime-python==2.0.13
#lightgbm>=3.3.1,<4.0
#matplotlib>=3.4.3,<3.5
#markupsafe>=2.1.1,<3.0
#nltk>=3.6.3,<4.0
#numpy>=1.22,<1.24
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
#pandas>=1.3.3,<1.6.0
#pandas-profiling>=3.1.0,<4.0
#Pillow>=9.3.0,<10.0
#pyarrow==4.0.1
#pyspark==3.5.2
pytesseract
pyYAML==5.4.1
#pytest>=6.2.4,<7.0
#scikit-learn>=1.0,<1.2.0
#scikit-optimize>=0.10.2,<0.11.0
#seaborn>=0.11.2,<0.12
#scikeras>=0.10.0,<0.11.0
#setuptools>=49.6.0
#shap==0.39.0
#six>=1.15.0
#smart-open>=6.0.0
#spacy>=3.0.0,<4.0
#statsmodels==0.13.5
#streamlit
#symspellpy>=6.7.8
#tensorflow>=2.8.0,<2.9.0
#tensorflow-estimator>=2.8.0,<2.9.0
textblob>=0.15.3,<0.16
#torch>=1.10.0,<1.11.0
#transformers>=4.21.0,<4.22.0
#typing-extensions>=3.7.4,<4.0.0
#vaderSentiment>=3.0.0
#wordcloud>=1.8.1,<2.0
#wrapt>=1.11.0,<1.15
#yfinance==0.2.50
#xgboost>=1.5.0,<1.7.0

